repost count issue fixed, comment like dislike apis done

+++ database/migrations/2022_11_15_225834_create_posts_table.php
@@ -35,9 +35,9 @@ return new class extends Migration
             $table->unsignedBigInteger('title_text_color_id')->nullable();
             $table->unsignedBigInteger('title_background_color_id')->nullable();
 
-            $table->smallInteger('likes')->default(0);
-            $table->smallInteger('repost')->default(0);
-            $table->smallInteger('shares')->default(0);
+            $table->unsignedBigInteger('likes')->default(0);
+            $table->unsignedBigInteger('repost')->default(0);
+            $table->unsignedBigInteger('shares')->default(0);
 
 
             $table->timestamps();
